cache the phantomjs binary file. exit on error code

+++ kolab-3.1-test/setup.sh
@@ -2,6 +2,11 @@ wget https://github.com/tpokorra/kolab3_tbits_scripts/archive/master.tar.gz
 tar xzf master.tar.gz
 cd kolab3_tbits_scripts-master/kolab3.1
 echo "y" | ./reinstall.sh
+if [ $? -ne 0 ]
+then
+  exit 1
+fi
+
 ./initSetupKolabPatches.sh
 setup-kolab --default --timezone=Europe/Berlin --directory-manager-pwd=test
 service kolabd restart
@@ -10,8 +15,22 @@ service kolabd restart
 ./initMailCatchall.sh
 
 # install python selenium for the tests
-yum -y install python-setuptools
+if [ -f /etc/centos-release ]
+then
+  yum -y install python-setuptools
+  cachepath=/var/cache/yum
+else
+  if [ -f /etc/lsb-release ]
+  then
+    apt-get install -y python-setuptools
+    cachepath=/var/cache/apt
+  fi
+fi
+
 easy_install selenium
-wget https://phantomjs.googlecode.com/files/phantomjs-1.9.2-linux-x86_64.tar.bz2
-tar xjf phantomjs-1.9.2-linux-x86_64.tar.bz
+if [ ! -f $cachepath/phantomjs-1.9.2-linux-x86_64.tar.bz2 ]
+then
+  wget -O $cachepath/phantomjs-1.9.2-linux-x86_64.tar.bz2 https://phantomjs.googlecode.com/files/phantomjs-1.9.2-linux-x86_64.tar.bz2
+fi
+tar xjf $cachepath/phantomjs-1.9.2-linux-x86_64.tar.bz2
 cp phantomjs-1.9.2-linux-x86_64/bin/phantomjs /usr/bin
